requirements.txt
aiogram~=2.14
environs~=8.0.0
asyncpg~=0.27.0
pandas~=1.5.2
python-dotenv~=0.21.1
telegraph~=2.2.0